Move models out of tests
