+++ test/nxos_ssh/test_getters.py
@@ -2,7 +2,7 @@
 
 from napalm.base.test.getters import BaseTestGetters, wrap_test_cases
 from napalm.base.test import helpers
-from napalm.base.test import models
+from napalm.base import models
 
 import pytest
 from mock import patch
